+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Typo in Student Guide feature Bus Pass [#609](https://github.com/rokwire/illinois-app/issues/609).
 
 ## [2.6.10] - 2021-07-26
 ### Changed